requirements.txt
fastapi==0.111.1
fasttext_wheel==0.9.2
pandas==2.2.2
pydantic==2.8.2
swifter==1.4.0
transformers==4.38.2
uvicorn==0.30.3
